+++ nestjs/src/core/decorators/index.ts
@@ -1,0 +1,1 @@
+export * from './student-id.decorator';
